[Refactor] finish rebuild userCenter each entry point
