+++ src/modules/userCenter/WaitReservation.tsx
@@ -1,0 +1,52 @@
+export default function WaitReservation() {
+  return (
+    <div className=" w-full rounded-2xl bg-bg2 text-center">
+      <ul className="flex w-full border-b border-secondary text-sm font-bold text-[#767494]">
+        <li className="w-[26.8421%] py-5 lg:w-[31.6205%]">諮商師</li>
+        <li className="w-[31.5789%] py-5 lg:w-[29.3478%]">諮商議題</li>
+        <li className="w-[41.57894%] py-5 lg:w-[39.0316%]">選擇時段</li>
+      </ul>
+
+      <ul className="mt-5 flex flex-col space-y-4 px-4 pb-9 text-sm text-primary-heavy lg:mt-7 lg:space-y-5 lg:px-7 lg:text-base">
+        <li className="flex items-center rounded-lg bg-white py-3 lg:py-[18px]">
+          <p className="w-[23.5632%] lg:w-[27.615%]">我真的</p>
+          <p className="w-[35.6321%] lg:w-[36.82%]">討厭table</p>
+          <div className="w-[40.8045%] lg:w-[35.5648%]">
+            <button
+              type="button"
+              className="rounded-full border border-primary-heavy py-1 px-[19px] text-xs font-semibold text-primary-heavy lg:py-2 lg:px-5 lg:text-sm lg:hover:bg-primary-heavy lg:hover:text-white"
+            >
+              選擇預約時段
+            </button>
+          </div>
+        </li>
+
+        <li className="flex items-center rounded-lg bg-white py-3 lg:py-[18px]">
+          <p className="w-[23.5632%] lg:w-[27.615%]">我真的</p>
+          <p className="w-[35.6321%] lg:w-[36.82%]">討厭table</p>
+          <div className="w-[40.8045%] lg:w-[35.5648%]">
+            <button
+              type="button"
+              className=" rounded-full border border-primary-heavy py-1 px-[19px] text-xs font-semibold text-primary-heavy lg:py-2 lg:px-5 lg:text-sm lg:hover:bg-primary-heavy lg:hover:text-white"
+            >
+              選擇預約時段
+            </button>
+          </div>
+        </li>
+
+        <li className="flex items-center rounded-lg bg-white py-3 lg:py-[18px]">
+          <p className="w-[23.5632%] lg:w-[27.615%]">我真的</p>
+          <p className="w-[35.6321%] lg:w-[36.82%]">討厭table</p>
+          <div className="w-[40.8045%] lg:w-[35.5648%]">
+            <button
+              type="button"
+              className=" rounded-full border border-primary-heavy py-1 px-[19px] text-xs font-semibold text-primary-heavy lg:py-2 lg:px-5 lg:text-sm lg:hover:bg-primary-heavy lg:hover:text-white"
+            >
+              選擇預約時段
+            </button>
+          </div>
+        </li>
+      </ul>
+    </div>
+  );
+}
